backend/requirements.txt
django
django-jsonview
django-redis
python-google-places
